requirements.txt
absl-py==2.1.0
astor==0.8.1
astunparse==1.6.3
autoflake==1.4
backcall==0.1.0
biopython==1.81
bleach==1.5.0
boto3==1.14.34
botocore==1.17.34
cachetools==4.0.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29.36
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
docutils==0.15.2
entrypoints==0.3
fastrlock==0.4
filelock==3.0.10
future==0.17.1
gast==0.3.3
gensim==3.8.3
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.10.1
grpcio==1.28.1
h5py==2.10.0
html5lib==0.9999999
hyperopt==0.2.3
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jmespath==0.10.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.1.0
matplotlib==3.7.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nose==1.3.7
notebook==5.7.4
numpy==1.26.4
oauthlib==3.1.0
onnx==1.16.0
opt-einsum==3.2.0
pandas==2.2.2
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
pydot==1.4.2
pyflakes==2.2.0
Pygments==2.3.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.6.0
PyYAML==3.13
pyzmq==25.1.2
qtconsole==4.4.3
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.10.0
semantic-version==2.6.0
Send2Trash==1.5.0
six==1.14.0
smart-open==6.4.0
tensorboard-plugin-wit==1.6.0.post2
tensorflow==2.17.1
tensorflow-estimator==2.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==6.3
tqdm==4.29.0
traitlets==4.3.2
typing-extensions==3.6.6
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
wrapt==1.12.1